+++ src/main/java/net/smokeybbq/bittermelon/medical/simulation/IVAdministration.java
@@ -7,7 +7,7 @@ import net.smokeybbq.bittermelon.medical.substance.Substance;
 public class IVAdministration extends PBPKModel {
     CirculatoryCompartment circulatory;
 
-    public IVAdministration(double dosage, Character character, Substance drug) {
+    public IVAdministration(float dosage, Character character, Substance drug) {
         super(dosage, character, drug);
     }
 
@@ -19,9 +19,9 @@ public class IVAdministration extends PBPKModel {
 
     @Override
     public void simulation() {
-        double circulatoryConcentration = circulatory.getConcentration(drug);
+        float circulatoryConcentration = circulatory.getConcentration(drug);
 
-        double circulatoryDerivative = circulatory.getDerivative(circulatoryConcentration, simpleCompartments, drug);
+        float circulatoryDerivative = circulatory.getDerivative(circulatoryConcentration, simpleCompartments, drug);
 
         handleSimpleCompartments();
         circulatory.updateConcentration(drug, circulatoryDerivative, timeStep);
