+++ src/main/java/net/smokeybbq/bittermelon/medical/simulation/OralAdministration.java
@@ -8,7 +8,7 @@ public class OralAdministration extends PBPKModel {
     EliminatingCompartment GI, liver;
     CirculatoryCompartment circulatory;
 
-    public OralAdministration(double dosage, Character character, Substance drug) {
+    public OralAdministration(float dosage, Character character, Substance drug) {
         super(dosage, character, drug);
     }
 
@@ -24,17 +24,17 @@ public class OralAdministration extends PBPKModel {
 
     @Override
     public void simulation() {
-        double GIConcentration = GI.getConcentration(drug);
-        double liverConcentration = liver.getConcentration(drug);
-        double circulatoryConcentration = circulatory.getConcentration(drug);
+        float GIConcentration = GI.getConcentration(drug);
+        float liverConcentration = liver.getConcentration(drug);
+        float circulatoryConcentration = circulatory.getConcentration(drug);
 
-        double liverBloodFlow = liver.getBloodFlow();
+        float liverBloodFlow = liver.getBloodFlow();
 
-        double GIDerivative = -drug.getAbsorptionRateConstant() * GIConcentration;
-        double liverDerivative = -GIDerivative - (drug.getMetabolismRateConstant() + liverBloodFlow) * liverConcentration + liverBloodFlow / 2 * circulatoryConcentration;
+        float GIDerivative = -drug.getAbsorptionRateConstant() * GIConcentration;
+        float liverDerivative = -GIDerivative - (drug.getMetabolismRateConstant() + liverBloodFlow) * liverConcentration + liverBloodFlow / 2 * circulatoryConcentration;
 
-        double circulatoryDrugSource = liverBloodFlow * liverConcentration;
-        double circulatoryDerivative = circulatory.getDerivative(circulatoryDrugSource, simpleCompartments, drug);
+        float circulatoryDrugSource = liverBloodFlow * liverConcentration;
+        float circulatoryDerivative = circulatory.getDerivative(circulatoryDrugSource, simpleCompartments, drug);
 
         handleSimpleCompartments();
