+++ src/main/java/net/smokeybbq/bittermelon/medical/substance/toxins/Toxin.java
@@ -3,12 +3,12 @@ package net.smokeybbq.bittermelon.medical.substance.toxins;
 import net.smokeybbq.bittermelon.medical.substance.Substance;
 
 public class Toxin extends Substance {
-    public Toxin(double absorptionModifier, double eliminationModifier, double metabolismModifier) {
+    public Toxin(float absorptionModifier, float eliminationModifier, float metabolismModifier) {
         super(absorptionModifier, eliminationModifier, metabolismModifier);
     }
 
     @Override
-    public void toxicDamage(double effectiveness) {
+    public void toxicDamage(float effectiveness) {
 
     }
 }
